Extract ButtonGroupLeft
