+++ packages/vscode-extension/src/webview/components/ButtonGroupLeft.css
@@ -1,0 +1,24 @@
+.button-group-left-wrapper {
+    position: absolute;
+    left: 0px;
+    top: 50%;
+    width: 50px;
+    height: 150px;
+    transform: translateY(-50%);
+    background: red;
+    overflow: visible;
+}
+
+.button-group-left-container {
+    display: flex;
+    position: absolute;
+    left: 10px;
+    top: 50%;
+    transform: translateY(-50%);
+}
+
+.button-group-left {
+    display: flex;
+    transition: 0.2s;
+    transform: translateX(-40px);
+}
